+++ include/graphics/entity.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace tomovis {
-
-class Entity {
-
-};
-
-} // namespace tomovis