+++ src/graphics/renderer.cpp
@@ -1,12 +1,11 @@
-#include <stdio.h>
 #include <algorithm>
+#include <stdio.h>
 
 #include <GL/gl3w.h>
 #include <GLFW/glfw3.h>
 #include <glm/gtx/transform.hpp>
 
 #include "graphics/renderer.hpp"
-
 
 namespace tomovis {
 
@@ -29,16 +28,14 @@ Renderer::Renderer() {
     glfwMakeContextCurrent(window_);
 
     gl3wInit();
-    glEnable(GL_MULTISAMPLE);  
+    glEnable(GL_MULTISAMPLE);
 
     glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 
     previous_time_ = glfwGetTime();
 }
 
-Renderer::~Renderer() {
-    glfwTerminate();
-}
+Renderer::~Renderer() { glfwTerminate(); }
 
 void Renderer::main_loop() {
     while (!glfwWindowShouldClose(window_)) {
@@ -51,17 +48,17 @@ void Renderer::main_loop() {
         const auto time_step = 0.0166666666;
         for (auto ticker : tickers_) {
             while (time_elapsed > time_step) {
-               ticker->tick(time_step);
-               time_elapsed -= time_step;
+                ticker->tick(time_step);
+                time_elapsed -= time_step;
             }
-           ticker->tick(time_elapsed);
+            ticker->tick(time_elapsed);
         }
 
         // Rendering
         int display_w, display_h;
         glfwGetFramebufferSize(window_, &display_w, &display_h);
         glViewport(0, 0, display_w, display_h);
-        glClearColor(1.0f, 1.0f, 1.0f, 1.0f);
+        glClearColor(0.1f, 0.1f, 0.1f, 1.0f);
         glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
         float ratio = (float)display_h / (float)display_w;
@@ -80,9 +77,7 @@ void Renderer::register_target(RenderTarget& target) {
     targets_.insert(&target);
 }
 
-void Renderer::register_ticker(Ticker& ticker) {
-    tickers_.push_back(&ticker);
-}
+void Renderer::register_ticker(Ticker& ticker) { tickers_.push_back(&ticker); }
 
 void Renderer::unregister_target(RenderTarget& target) {
     targets_.erase(std::find(targets_.begin(), targets_.end(), &target));