+++ src/graphics/scene_object.cpp
@@ -25,12 +25,12 @@ void SceneObject::tick(float time_elapsed) {
 }
 
 void SceneObject::describe() {
-    if (ImGui::CollapsingHeader("Camera", nullptr, ImGuiTreeNodeFlags_DefaultOpen)) {
+    if (ImGui::CollapsingHeader("camera")) {
         camera_->describe();
     }
 
     for (auto& id_and_comp : components_) {
-        if (ImGui::CollapsingHeader(id_and_comp.first.c_str(), nullptr, ImGuiTreeNodeFlags_DefaultOpen)) {
+        if (ImGui::CollapsingHeader(id_and_comp.first.c_str())) {
             id_and_comp.second->describe();
         }
     }